src/main/java/com/sysone/compress/exception/RestExceptionHandler.java
package com.sysone.compress.exception;

import com.sysone.compress.dto.ErrorMessage;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.ObjectError;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.client.HttpStatusCodeException;
import org.springframework.web.context.request.WebRequest;
import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;

import java.util.ArrayList;
import java.util.List;

import static com.sysone.compress.constant.Constants.VALIDATE_FAILED;

@ControllerAdvice
public class RestExceptionHandler extends ResponseEntityExceptionHandler {

    public RestExceptionHandler() {
        super();
    }

    @ExceptionHandler(HttpStatusCodeException.class)
    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex,
              HttpHeaders headers, HttpStatus status, WebRequest request) {
        ErrorMessage errorMessage = new ErrorMessage();
        errorMessage.setError(VALIDATE_FAILED + getMessageErrors(ex.getBindingResult().getAllErrors()));
        return handleExceptionInternal(ex, errorMessage, new HttpHeaders(), HttpStatus.BAD_REQUEST, request);
    }

    private List<String> getMessageErrors(List<ObjectError> errors) {
        List<String> messageErrors = new ArrayList<>();
        for(ObjectError objectError : errors) {
            messageErrors.add(objectError.getDefaultMessage());
        }
        return messageErrors;
    }

}
